requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
bitarray==2.8.2
blake3==0.3.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cytoolz==0.12.2
distlib==0.3.7
editables==0.5
eth-abi==4.2.1
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.0
eth-utils==2.2.2
farcaster @ file:///Users/vrypan/Devel/farcaster-py
filelock==3.12.4
frozenlist==1.4.0
grpcio==1.59.0
grpcio-tools==1.59.0
h11==0.14.0
hatch==1.7.0
hatchling==1.18.0
hexbytes==0.3.1
httpcore==0.18.0
httpx==0.25.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
jaraco.classes==3.3.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
keyring==24.2.0
lru-dict==1.2.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.1.0
multidict==6.0.4
packaging==23.2
parsimonious==0.9.0
pathspec==0.11.2
pexpect==4.8.0
platformdirs==3.11.0
pluggy==1.3.0
protobuf==4.24.4
ptyprocess==0.7.0
pycparser==2.21
pycryptodome==3.19.0
Pygments==2.16.1
PyNaCl==1.5.0
pyperclip==1.8.2
python-dotenv==1.0.0
pyunormalize==15.0.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rlp==3.0.0
rpds-py==0.10.6
shellingham==1.5.3
sniffio==1.3.0
tomli_w==1.0.0
tomlkit==0.12.1
toolz==0.12.0
trove-classifiers==2023.10.18
typing_extensions==4.8.0
urllib3==2.0.7
userpath==1.9.1
virtualenv==20.24.5
web3==6.11.1
websockets==11.0.3
yarl==1.9.2
zipp==3.17.0
